ref(compiler/ssa)[sty]: move all SSA stuff in a separate crate
