+++ compiler/tidec_codegen_llvm/Cargo.toml
@@ -9,7 +9,6 @@ inkwell = { version = "0.5.0", features = ["llvm18-0"] } # inkwell = { git = "ht
 tidec_abi = { path = "../tidec_abi" }
 tidec_codegen_ssa = { path = "../tidec_codegen_ssa" }
 tidec_lir = { path = "../tidec_lir" }
-tidec_utils = { path = "../tidec_utils" }
 tracing = "0.1.41"
 # tidy-alphabetical-end
