+++ compiler/tidec_codegen_llvm/src/builder.rs
@@ -4,9 +4,7 @@ use inkwell::types::BasicType;
 use inkwell::values::{AnyValueEnum, FunctionValue};
 use inkwell::{basic_block::BasicBlock, builder::Builder};
 use tidec_abi::TyAndLayout;
-use tidec_codegen_ssa::traits::{
-    BuilderMethods, CodegenBackend, CodegenBackendTypes, CodegenMethods, PreDefineCodegenMethods,
-};
+use tidec_codegen_ssa::traits::{BuilderMethods, CodegenBackendTypes};
 use tidec_lir::syntax::LirTy;
 use tracing::instrument;
