+++ compiler/tidec_codegen_ssa/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2024"
 # tidy-alphabetical-start
 tidec_abi = { path = "../tidec_abi" }
 tidec_lir = { path = "../tidec_lir" }
+tidec_utils = { path = "../tidec_utils" }
 tracing = "0.1.41"
 # tidy-alphabetical-end
