+++ compiler/tidec_codegen_ssa/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod traits;
+pub mod entry;
+pub mod lir;
